fix: reimplemented the login design
