--- src/pages/UserPage/userPagess.module.css
@@ -1,328 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Nunito&display=swap');
-.user_page_container {
-	padding: 2.5625rem 4rem;
-	width: 100vw;
-	max-width: 100vw;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;	
-}
-.user_header_container {
-	display: flex;
-	width: 100%;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-}
-
-.user_page_heading {
-	font-family: 'Nunito';
-	font-style: normal;
-	font-weight: 600;
-	font-size: 48px;
-	width: 50%;
-	text-align: center;
-	line-height: -0.3px;
-	color: #492B7C;
-
-}
-
-/* search box component style */
-.search_container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	gap: 4rem;
-	margin: 4rem 0;
-}
-.searchbox_container {
-	width: 100%;
-	margin: 0 auto;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: 5px 25px;
-	height: 52px;
-}
-.search_box {
-	width: 594px;
-	height: 100%;
-	outline: none;
-	padding: 5px 25px;
-	border: 0.68003px solid #b4b4b4;
-	border-radius: 21.761px 0px 0px 21.761px;
-}
-.search_languages {
-	display: flex;
-	padding: 5px 27px;
-	height: 100%;
-	gap: 21px;
-	justify-content: center;
-	align-items: center;
-	border: 0.68px solid #b4b4b4;
-	border-radius: 0px 22px 22px 0px;
-}
-
-.search_languages > button {
-	background-color: white;
-	border: 0px;
-}
-
-.search_languages > button:hover {
-	background-color: #1D1DD8;
-	color: white;
-	padding: 7px;
-	border-radius: 10px;
-}
-.follow_button {
-	background-color: white;
-	color: #1D1DD8 ;
-	padding: 10px 9px 10px 9px;
-	border :1px solid #1D1DD8;
-	border-radius: 5px;
-}
-.follow_text {
-	color: #8e8eec;
-}
-/* user profile card component styles */
-.user_profile_card {
-	width: 100%;
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	background: #fefefe;
-	border: 0.152023px solid #bababa;
-	box-shadow: 0.760116px 2.28035px 3.04046px rgba(0, 0, 0, 0.15);
-	border-radius: 4.56069px;
-	padding: 10px 10px 10px 10px;
-}
-
-.user_profile_image {
-	width: 60px;
-	height: 60px;
-	display: block;
-	border-radius: 100%;
-	margin-bottom: 72px;
-}
-
-.user_info_container {
-	width: 100%;
-	justify-content: space-between;
-	height: 140px;
-	display: flex;
-	flex-direction: column;
-	padding: 10px 10px 10px 10px;
-	
-	gap: 10px;
-}
-
-.user_follow_info {
-	display: flex;
-	gap: 15px;
-    align-items: center;
-}
-
-.user_name {
-	font-weight: 700;
-	font-size: 1.5rem;
-	line-height: 18px;
-}
-
-.user_stack {
-	font-size: 14px;
-	line-height: 13px;
-	font-weight: 400;
-	margin-top: 0.75rem;
-}
-
-.user_extra_info {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-}
-
-.user_followers {
-	font-weight: 500;
-	font-size: 12px;
-	color: #bababa;
-}
-
-.user_favorite_language {
-	font-size: 12px;
-	line-height: 10px;
-	color: #8e8eec;
-}
-.user_grid_container {
-	display: grid;
-	justify-content: center;
-	align-items: center;
-	padding: 1rem;
-	justify-content: center;
-	place-items: center;
-	grid-template-columns: repeat(3, calc(73% / 3 - 20px));
-	gap: 2rem;
-
-}
-
-/* user filter bottom style */
-.users_filter {
-	width: 100vw;
-	padding: 2rem 4rem;
-	justify-content: flex-end;
-	display: flex;
-	align-items: center;
-	margin-right: 130px;
-}
-.filter_options {
-	display: flex;
-	gap: 1rem;
-	align-items: center;
-}
-.filter_options > button {
-	padding: 6px;
-	text-align: center;
-	outline: none;
-	border: 0.485714px solid #b4b4b4;
-}
-.more_search {
-	display: flex;
-	gap: 1.5rem;
-	align-items: center;
-}
-.more_search > a {
-	color: black;
-	text-decoration: none;
-}
-.more_search > a:hover {
-	color: black;
-	border-bottom: 2px solid #1d1dd8;
-}
-.more_search > button {
-	outline: none;
-	border: none;
-	width: 76.85px;
-	text-align: center;
-	color: white;
-	height: 26.29px;
-	background: #1d1dd8;
-	border-radius: 3.06434px;
-}
-.select_lang {
-	display: none;
-}
-
-@media screen and (max-width: 1000px) {
-	.searchbox_container {
-		width: 100vw;
-	}
-	.search_box {
-		width: 55%;
-	}
-	.search_languages {
-		width: 45%;
-	}
-	.users_filter {
-		margin-right: 0px;
-	}
-}
-@media screen and (max-width: 850px) {
-	.user_profile_image {
-		width: 30%;
-		max-width: 76px;
-		display: block;
-		border-radius: 100%;
-	}
-	.users_filter {
-		margin-right: 0px;
-	}
-}
-@media screen and (max-width: 1200px) {
-	.user_page_container {
-		padding: 2rem;
-	}
-	.user_header_container {
-		gap: 1rem;
-	}
-	.user_page_heading {
-		width: 100%;
-		font-size: 1.5rem;
-	}
-	.search_container {
-		margin: 2rem 0;
-		gap: 2rem;
-	}
-	.searchbox_container {
-		align-items: center;
-		flex-direction: row;
-		width: 80vw;
-		margin: 0 auto;
-	}
-
-	.search_languages {
-		display: none;
-	}
-	.search_box {
-		width: 70%;
-	}
-	.select_lang {
-		display: block;
-		height: 100%;
-		width: 30%;
-		padding: 0.5rem;
-		border: 0.68px solid #b4b4b4;
-		border-radius: 0px 22px 22px 0px;
-	}
-	.users_filter {
-		justify-content: center;
-	}
-	.filter_hidden {
-		display: none;
-	}
-	.user_profile_card {
-		width: 335px;
-	}
-	.user_profile_image {
-		width: 60px;
-		height: 60px;
-	}
-	.user_info_container {
-		width: 250px;
-	}
-	.user_name {
-		font-size: 1.2rem;
-	}
-	.user_favorite_language {
-		word-wrap: break-word;
-	}
-	.user_grid_container {
-		grid-template-columns: repeat(auto-fill, minmax(335px, 1fr));
-		gap: 2rem;
-	}	
-	.users_filter {
-		margin-right: 0px;
-	}
-}
-
-@media screen and (max-width: 340px) {
-	.searchbox_container {
-		width: 90vw;
-	}
-	.user_info_container {
-		width: 235px;
-	}
-	.user_profile_card {
-		justify-content: flex-start;
-		gap: 1rem;
-	}
-	.user_grid_container {
-		grid-template-columns: repeat(auto-fill, minmax(330px, 1fr));
-		gap: 2rem;
-	}
-	.users_filter {
-		margin-right: 0px;
-	}
-}
